+++ modules/tinymce/src/plugins/link/test/ts/browser/RemoveLinkTest.ts
@@ -1,4 +1,4 @@
-import { describe, it } from '@ephox/bedrock-client';
+import { context, describe, it } from '@ephox/bedrock-client';
 import { TinyAssertions, TinyHooks, TinySelections, TinyUiActions } from '@ephox/wrap-mcagar';
 
 import Editor from 'tinymce/core/api/Editor';
@@ -8,7 +8,8 @@ describe('browser.tinymce.plugins.link.RemoveLinkTest', () => {
   const hook = TinyHooks.bddSetupLight<Editor>({
     plugins: 'link',
     toolbar: 'unlink',
-    base_url: '/project/tinymce/js/tinymce'
+    base_url: '/project/tinymce/js/tinymce',
+    indent: false
   }, [ Plugin ]);
 
   it('TBA: Removing a link with a collapsed selection', async () => {
@@ -45,5 +46,25 @@ describe('browser.tinymce.plugins.link.RemoveLinkTest', () => {
     TinyUiActions.clickOnToolbar(editor, 'button[title="Remove link"]');
     TinyAssertions.assertContentPresence(editor, { a: 0 });
     TinyAssertions.assertSelection(editor, [ 0, 0 ], 1, [ 0, 4 ], 2);
+  });
+
+  context('Block links', () => {
+    it('TINY-9172: Removing root level link should convert it to regular text block', () => {
+      const editor = hook.editor();
+      editor.setContent('<a href="#"><p>tiny</p></a>');
+      TinySelections.setCursor(editor, [ 0, 0, 0 ], 1);
+      TinyUiActions.clickOnToolbar(editor, 'button[title="Remove link"]');
+      TinyAssertions.assertContent(editor, '<p>tiny</p>');
+      TinyAssertions.assertCursor(editor, [ 0, 0 ], 1);
+    });
+
+    it('TINY-9172: Removing wrapped block link should unwrap', () => {
+      const editor = hook.editor();
+      editor.setContent('<div><a href="#"><p>tiny</p></a></div>');
+      TinySelections.setCursor(editor, [ 0, 0, 0, 0 ], 1);
+      TinyUiActions.clickOnToolbar(editor, 'button[title="Remove link"]');
+      TinyAssertions.assertContent(editor, '<div><p>tiny</p></div>');
+      TinyAssertions.assertCursor(editor, [ 0, 0, 0 ], 1);
+    });
   });
 });