+++ modules/tinymce/src/themes/silver/main/ts/backstage/ColorInputBackstage.ts
@@ -8,8 +8,8 @@ import * as Options from '../ui/core/color/Options';
 export interface UiFactoryBackstageForColorInput {
   readonly colorPicker: (callback: ColorSwatch.ColorInputCallback, value: string) => void;
   readonly hasCustomColors: () => boolean;
-  readonly getColors: () => Menu.ChoiceMenuItemSpec[];
-  readonly getColorCols: () => number;
+  readonly getColors: (id: string) => Menu.ChoiceMenuItemSpec[];
+  readonly getColorCols: (id: string) => number;
 }
 
 const colorPicker = (editor: Editor) => (callback: ColorSwatch.ColorInputCallback, value: string) => {
@@ -19,9 +19,9 @@ const colorPicker = (editor: Editor) => (callback: ColorSwatch.ColorInputCallbac
 
 const hasCustomColors = (editor: Editor) => (): boolean => Options.hasCustomColors(editor);
 
-const getColors = (editor: Editor) => (): Menu.ChoiceMenuItemSpec[] => Options.getColors(editor);
+const getColors = (editor: Editor) => (id: string): Menu.ChoiceMenuItemSpec[] => Options.getColors(editor, id);
 
-const getColorCols = (editor: Editor) => (): number => Options.getColorCols(editor);
+const getColorCols = (editor: Editor) => (id: string): number => Options.getColorCols(editor, id);
 
 export const ColorInputBackstage = (editor: Editor): UiFactoryBackstageForColorInput => ({
   colorPicker: colorPicker(editor),