+++ modules/tinymce/src/themes/silver/test/ts/headless/components/colorinput/ColorInputTest.ts
@@ -23,6 +23,7 @@ describe('headless.tinymce.themes.silver.components.colorinput.ColorInputTest', 
       components: [
         renderColorInput({
           name: 'alpha',
+          storageKey: 'test',
           label: Optional.some('test-color-input')
         }, helpers.shared(), {
           colorPicker: (_callback, _value) => {},