+++ modules/bridge/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- Added optional `storageKey` property to colorinput component #TINY-9184
+- New `view` component. #TINY-9210
+
 ### Fixed
 - Fixed type errors caused by upgrading to TypeScript 4.8. #TINY-9161